rename UI components
